--- springdoc-openapi-webmvc-core/src/test/java/test/org/springdoc/api/app69/SpringDocApp69Test.java
@@ -1,6 +1,0 @@
-package test.org.springdoc.api.app69;
-
-import test.org.springdoc.api.AbstractSpringDocTest;
-
-public class SpringDocApp69Test extends AbstractSpringDocTest {
-}
