+++ springdoc-openapi-webmvc-core/src/test/java/test/org/springdoc/api/app70/SpringDocApp70Test.java
@@ -1,0 +1,6 @@
+package test.org.springdoc.api.app70;
+
+import test.org.springdoc.api.AbstractSpringDocTest;
+
+public class SpringDocApp70Test extends AbstractSpringDocTest {
+}
