+++ springdoc-openapi-webmvc-core/src/test/java/test/org/springdoc/api/app70/customizer/CustomizedOperation.java
@@ -1,4 +1,4 @@
-package test.org.springdoc.api.app69.customizer;
+package test.org.springdoc.api.app70.customizer;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
