+++ springdoc-openapi-webmvc-core/src/test/resources/results/app70.json
